Backend/requirements.txt
fastapi
uvicorn
motor
python-dotenv
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
pymongo
python-multipart
email-validator
pillow
torch
transformers
accelerate
bitsandbytes
requests
httpx>=0.24.0
pytest
pytest-asyncio
